requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
apscheduler==3.10.4
jinja2==3.1.4
aiofiles==24.1.0
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.1.0
